refactor: Spin the platform
It's now possible to spin and compute the load, but it's not possible to spin 1 billion times yet.

Plotting the load after each spin cycle, though, it is clear the example data is quickly reaching some kind of limit cycle, i.e. the load oscillates.

The load for the actual data seems to converge to something near 95 k, but need to investigate further